Add GetProductDetails API to ItemsController
Added a new HTTP GET action method `GetProductDetails` to the
`ItemsController` in the `ShoppingCartApp.Controllers` namespace.
This method retrieves product details asynchronously from the
database using `_context.Products.FindAsync(id)`. If the product
is not found, it returns a `NotFound()` response. Otherwise, it
returns a JSON object containing the product's name, price, and
image. This endpoint enables dynamic fetching of product details.